Reparenting was not working because animations weren't triggering.
In the deuxnius prototype, when transitioning between pager pages, the
focus was being destroyed which caused the focus ring to go to display:
none.  Then when the new page took effect, the focus ring was
transitioning from display: none to real and because we were setting the
new coordinates at the same time without a reflow, no transition was
happening.  As a result, no reparenting occurred when it got to its
target.  So now we force a reflow.