add a checkbox binding, introduce support for JS attr bindings for that
Specifically, the DOM attribute "checked" for checkbox "input" elements
has weird semantics that define the initial/default state of the
checkbox by virtue of being present-at-all rather than having its value
matter.  It is rather the JS/IDL attribute "checked" on the DOM node
that is the live value and what we want to mess with.

Our click-to-focus logic is still being blind to the focusability of
form items, which is annoying and about to be rectified.