labelTransformer was not being used; provide key to the value transformer
In ArbPL it is handy to be able to transform things that look like they
might be millisecond timestamps so that they also include a human readable
version of the date along-side.  We accomplish this by telling the value
transformer the key that was the origin of the value, since a Number on its
own obviously lacks useful type info.